initial swap working
