+++ crates/codegen/src/python.rs
@@ -322,122 +322,4 @@ def even_fib(n):
             // println!("{}", code);
         }
     }
-
-    pub fn make_floating_point_add() -> CFG {
-        let code = r#"
-def floating_point_add(a_sign, a_exponent, a_mantissa, b_sign, b_exponent, b_mantissa):
-    # Make sure a has larger by magnitude
-    if a_exponent < b_exponent:
-        temp_sign = a_sign
-        a_sign = b_sign
-        b_sign = temp_sign
-
-        temp_exponent = a_exponent
-        a_exponent = b_exponent
-        b_exponent = temp_exponent
-
-        temp_mantissa = a_mantissa
-        a_mantissa = b_mantissa
-        b_mantissa = temp_mantissa
-
-    elif a_exponent == b_exponent:
-        if a_mantissa < b_mantissa:
-            temp_sign = a_sign
-            a_sign = b_sign
-            b_sign = temp_sign
-
-            temp_exponent = a_exponent
-            a_exponent = b_exponent
-            b_exponent = temp_exponent
-
-            temp_mantissa = a_mantissa
-            a_mantissa = b_mantissa
-            b_mantissa = temp_mantissa
-
-
-    c_sign = a_sign
-
-    # Add implicit one
-    a_mantissa |= 1 << 23
-    b_mantissa |= 1 << 23
-
-    # Adjust the smaller mantissa so exponents are same
-    exponent_difference = a_exponent - b_exponent
-    b_mantissa >>= exponent_difference
-
-    subtract = a_sign ^ b_sign
-
-    if subtract:
-        c_mantissa = a_mantissa - b_mantissa
-    else:
-        c_mantissa = a_mantissa + b_mantissa
-
-    c_exponent = a_exponent
-
-    # Normalize
-    msb_index = 0
-    temp = c_mantissa
-
-    while temp:
-        # Logical shift right (Python only does arithmetic)
-        if temp >= 0:
-            temp >>= 1
-        else:
-            temp = (temp + (1 << 24)) >> 1
-
-        msb_index += 1
-
-    # Shift left until implicit bit is MSB
-    # Decrease exponent to match
-    left_shift_amount = 24 - msb_index  # Can be negative
-    if left_shift_amount >= 0:
-        c_mantissa <<= left_shift_amount
-        c_exponent -= left_shift_amount
-    else:
-        c_mantissa >>= -left_shift_amount
-        c_exponent += -left_shift_amount
-
-    c_mantissa &= (1 << 23) - 1
-
-    # return c_sign, c_exponent, c_mantissa
-    return c_mantissa
-"#;
-        let visitor = tohdl_frontend::AstVisitor::from_text(code);
-
-        visitor.get_graph()
-    }
-
-    #[test]
-    fn floating_point_add() {
-        let mut graph = make_floating_point_add();
-
-        let mut manager = PassManager::default();
-
-        manager.add_pass(InsertFuncNodes::transform);
-        manager.add_pass(InsertCallNodes::transform);
-        manager.add_pass(BraunEtAl::transform);
-        // manager.add_pass(InsertPhi::transform);
-        // manager.add_pass(MakeSSA::transform);
-        // manager.add_pass(RemoveUnreadVars::transform);
-        // manager.add_pass(RemoveRedundantCalls::transform);
-
-        manager.apply(&mut graph);
-
-        let mut lower = tohdl_passes::transform::LowerToFsm::default();
-        lower.apply(&mut graph);
-
-        graph.write_dot("graph.dot");
-
-        // println!("original to subgraph {:?}", lower.node_to_subgraph);
-
-        // Write all new subgraphs to files
-        for (i, subgraph) in lower.get_subgraphs().iter().enumerate() {
-            subgraph.write_dot(format!("lower_to_fsm_{}.dot", i).as_str());
-
-            let mut codegen = CodeGen::new(subgraph.clone(), i, lower.get_external_funcs(i));
-            codegen.work(subgraph.get_entry());
-            let code = codegen.get_code();
-            // println!("{}", code);
-        }
-    }
 }
