+++ crates/pytohdl/src/lib.rs
@@ -1,7 +1,5 @@
 use pyo3::prelude::*;
-use tohdl_codegen::verilog::{
-    graph_to_verilog, Context,
-};
+use tohdl_codegen::verilog::{graph_to_verilog, Context};
 
 /// Formats the sum of two numbers as string.
 #[pyfunction]
@@ -33,4 +31,61 @@ fn translate(code: &str) -> String {
     let visitor = tohdl_frontend::AstVisitor::from_text(code);
     let graph = visitor.get_graph();
     graph_to_verilog(graph)
+}
+
+mod tests {
+    use super::*;
+
+    #[test]
+    pub fn floating_point_add() {
+        let code = r#"
+def floating_point_add(a_sign, a_exponent, a_mantissa, b_sign, b_exponent, b_mantissa):
+    # Make sure a has larger by magnitude
+    if a_exponent < b_exponent:
+        temp_sign = a_sign
+        a_sign = b_sign
+        b_sign = temp_sign
+
+        temp_exponent = a_exponent
+        a_exponent = b_exponent
+        b_exponent = temp_exponent
+
+        temp_mantissa = a_mantissa
+        a_mantissa = b_mantissa
+        b_mantissa = temp_mantissa
+
+    elif a_exponent == b_exponent:
+        if a_mantissa < b_mantissa:
+            temp_sign = a_sign
+            a_sign = b_sign
+            b_sign = temp_sign
+
+            temp_exponent = a_exponent
+            a_exponent = b_exponent
+            b_exponent = temp_exponent
+
+            temp_mantissa = a_mantissa
+            a_mantissa = b_mantissa
+            b_mantissa = temp_mantissa
+
+    c_sign = a_sign
+
+    # Add implicit one
+    a_mantissa |= 1 << 23
+    b_mantissa |= 1 << 23
+
+    yield c_sign
+    yield c_sign
+    yield a_mantissa
+        "#;
+        let visitor = tohdl_frontend::AstVisitor::from_text(code);
+        let graph = visitor.get_graph();
+        let verilog = graph_to_verilog(graph);
+
+        use std::fs::File;
+        use std::io::Write;
+
+        let mut file = File::create("output.sv").unwrap();
+        write!(file, "{}", verilog).unwrap();
+    }
 }
